refactor GenerateRespose
