+++ packages/mongodb-chatbot-server/src/processors/includeChunksForMaxTokensPossible.ts
@@ -1,0 +1,19 @@
+import { EmbeddedContent } from "mongodb-rag-core";
+
+/**
+      This function returns the chunks that can fit in the maxTokens.
+      It limits the number of tokens that are sent to the LLM.
+      */
+export function includeChunksForMaxTokensPossible({
+  maxTokens,
+  content,
+}: {
+  maxTokens: number;
+  content: EmbeddedContent[];
+}): EmbeddedContent[] {
+  let total = 0;
+  const fitRangeEndIndex = content.findIndex(
+    ({ tokenCount }) => (total += tokenCount) > maxTokens
+  );
+  return fitRangeEndIndex === -1 ? content : content.slice(0, fitRangeEndIndex);
+}
