+++ packages/mongodb-chatbot-server/src/processors/index.ts
@@ -5,5 +5,5 @@ export * from "./QueryPreprocessorFunc";
 export * from "./filterOnlySystemPrompt";
 export * from "./makeDefaultReferenceLinks";
 export * from "./makeFilterNPreviousMessages";
-export * from "./makeRagGenerateUserPrompt";
 export * from "./makeVerifiedAnswerGenerateUserPrompt";
+export * from "./includeChunksForMaxTokensPossible";
