+++ packages/mongodb-chatbot-server/src/routes/conversations/addMessageToConversation.ts
@@ -16,7 +16,6 @@ import {
   Conversation,
   SomeMessage,
   makeDataStreamer,
-  ChatLlm,
 } from "mongodb-rag-core";
 import {
   ApiMessage,
@@ -31,12 +30,9 @@ import {
   AddCustomDataFunc,
   ConversationsRouterLocals,
 } from "./conversationsRouter";
-import { GenerateUserPromptFunc } from "../../processors/GenerateUserPromptFunc";
-import { FilterPreviousMessages } from "../../processors/FilterPreviousMessages";
-import { filterOnlySystemPrompt } from "../../processors/filterOnlySystemPrompt";
-import { generateResponse, GenerateResponseParams } from "../generateResponse";
 import { wrapTraced } from "mongodb-rag-core/braintrust";
 import { UpdateTraceFunc, updateTraceIfExists } from "./UpdateTraceFunc";
+import { GenerateResponse, GenerateResponseParams } from "../GenerateResponse";
 
 export const DEFAULT_MAX_INPUT_LENGTH = 3000; // magic number for max input size for LLM
 export const DEFAULT_MAX_USER_MESSAGES_IN_CONVERSATION = 7; // magic number for max messages in a conversation
@@ -66,11 +62,9 @@ export type AddMessageRequest = z.infer<typeof AddMessageRequest>;
 
 export interface AddMessageToConversationRouteParams {
   conversations: ConversationsService;
-  llm: ChatLlm;
-  generateUserPrompt?: GenerateUserPromptFunc;
-  filterPreviousMessages?: FilterPreviousMessages;
   maxInputLengthCharacters?: number;
   maxUserMessagesInConversation?: number;
+  generateResponse: GenerateResponse;
   addMessageToConversationCustomData?: AddCustomDataFunc;
   /**
     If present, the route will create a new conversation
@@ -86,11 +80,6 @@ export interface AddMessageToConversationRouteParams {
       when it is created.
      */
     addCustomData?: AddCustomDataFunc;
-    /**
-      The system message to add to the new conversation
-      when it is created.
-     */
-    systemMessage?: SystemMessage;
   };
 
   /**
@@ -114,11 +103,9 @@ type MakeTracedResponseParams = Pick<
 
 export function makeAddMessageToConversationRoute({
   conversations,
-  llm,
-  generateUserPrompt,
+  generateResponse,
   maxInputLengthCharacters = DEFAULT_MAX_INPUT_LENGTH,
   maxUserMessagesInConversation = DEFAULT_MAX_USER_MESSAGES_IN_CONVERSATION,
-  filterPreviousMessages = filterOnlySystemPrompt,
   addMessageToConversationCustomData,
   createConversation,
   updateTrace,
@@ -150,14 +137,7 @@ export function makeAddMessageToConversationRoute({
           dataStreamer,
           shouldStream,
           reqId,
-          llm,
           conversation,
-          generateUserPrompt,
-          filterPreviousMessages,
-          llmNotWorkingMessage:
-            conversations.conversationConstants.LLM_NOT_WORKING,
-          noRelevantContentMessage:
-            conversations.conversationConstants.NO_RELEVANT_CONTENT,
         });
       },
       {
@@ -425,9 +405,6 @@ const loadConversation = async ({
       message: stripIndents`Creating new conversation`,
     });
     return await conversations.create({
-      initialMessages: createConversation.systemMessage
-        ? [createConversation.systemMessage]
-        : undefined,
       customData: createConversation.addCustomData
         ? await createConversation.addCustomData(req, res)
         : undefined,
