+++ packages/mongodb-chatbot-server/src/routes/conversations/conversationsRouter.ts
@@ -35,6 +35,7 @@ import {
   makeGetConversationRoute,
 } from "./getConversation";
 import { UpdateTraceFunc } from "./UpdateTraceFunc";
+import { GenerateResponse } from "../GenerateResponse";
 
 /**
   Configuration for rate limiting on the /conversations/* routes.
@@ -118,30 +119,17 @@ export type ConversationsMiddleware = RequestHandler<
   Configuration for the /conversations/* routes.
  */
 export interface ConversationsRouterParams {
-  llm: ChatLlm;
   conversations: ConversationsService;
-  systemPrompt: SystemPrompt;
-
   /**
-    Function to generate the user prompt sent to the {@link ChatLlm}.
-    You can perform any preprocessing of the user's message
-    including retrieval augmented generation here.
+    Logic to generate the response on the addMessageToConversation route.
    */
-  generateUserPrompt?: GenerateUserPromptFunc;
+  generateResponse: GenerateResponse;
 
   /**
     Maximum number of characters in user input.
     Server returns 400 error if user input is longer than this.
    */
   maxInputLengthCharacters?: number;
-
-  /**
-    Function to filter which previous messages are sent to the {@link ChatLlm}.
-    For example, you may only want to send the system prompt to the LLM
-    with the user message or the system prompt and X prior messages.
-    Defaults to sending only the system prompt.
-   */
-  filterPreviousMessages?: FilterPreviousMessages;
 
   /**
     Maximum number of user-sent messages in a conversation.
@@ -267,14 +255,11 @@ export const defaultAddMessageToConversationCustomData: AddDefinedCustomDataFunc
   Constructor function to make the /conversations/* Express.js router.
  */
 export function makeConversationsRouter({
-  llm,
   conversations,
-  systemPrompt,
+  generateResponse,
   maxInputLengthCharacters,
   maxUserMessagesInConversation,
-  filterPreviousMessages,
   rateLimitConfig,
-  generateUserPrompt,
   middleware = [requireValidIpAddress(), requireRequestOrigin()],
   createConversationCustomData = defaultCreateConversationCustomData,
   addMessageToConversationCustomData = defaultAddMessageToConversationCustomData,
@@ -329,7 +314,6 @@ export function makeConversationsRouter({
     makeCreateConversationRoute({
       conversations,
       createConversationCustomData,
-      systemPrompt,
     })
   );
 
@@ -364,20 +348,17 @@ export function makeConversationsRouter({
    */
   const addMessageToConversationRoute = makeAddMessageToConversationRoute({
     conversations,
-    llm,
     maxInputLengthCharacters,
     maxUserMessagesInConversation,
     addMessageToConversationCustomData,
-    generateUserPrompt,
-    filterPreviousMessages,
     createConversation: createConversationOnNullMessageId
       ? {
           createOnNullConversationId: createConversationOnNullMessageId,
           addCustomData: createConversationCustomData,
-          systemMessage: systemPrompt,
         }
       : undefined,
     updateTrace: addMessageToConversationUpdateTrace,
+    generateResponse,
   });
   conversationsRouter.post(
     "/:conversationId/messages",
